third_party/blink/web_tests/fast/parser/entity-end-script-tag-expected.txt

CONSOLE ERROR: line 11: Uncaught SyntaxError: Unexpected token '&'
Test parsing of entity-escaped </script> tag for Bug 7931: Escaped elements within a textarea block can cause the textarea box to be closed prematurely

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Got expected error: 'SyntaxError'
PASS document.getElementById("myscript").textContent is expectedResult
PASS successfullyParsed is true

TEST COMPLETE
